feat: better version managing wasplib
Doesn't use git2 client anymore.
It simply deletes the folder, downloads the zip if it hasn't already and unzips it.
For this use case this is 20 times faster and for this speed and reliability is all that matters.
In the future, I'll change wasp-plugins to be the same as speed is important for this